% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_INIT.R
\name{print.INIT}
\alias{print.INIT}
\title{Print function for INIT objects}
\usage{
\method{print}{INIT}(x, ...)
}
\arguments{
\item{x}{A INIT object}

\item{...}{}
}
\description{
Print function for INIT objects
}
