\name{print.INIT}
\alias{print.INIT}
\alias{summary.INIT}
\title{
Print method for INIT tests
}
\description{
Prints a short overview of the results of \code{\link{INIT}}
}
\usage{
\method{print}{INIT}(x, ...)
\method{summary}{INIT}(object, ...)
}
\arguments{
  \item{x}{
A \code{INIT} object
}
  \item{object}{
A \code{INIT} object
}
  \item{\dots}{
Not used.
}
}