% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INIT.R
\name{INIT}
\alias{INIT}
\title{Function to test for homogeneity between idiographic network models}
\usage{
INIT(
  data,
  vars,
  dayvar,
  beepvar,
  idvar,
  estimator = "ML",
  network_type = "saturated",
  homogeneity_test = "overall",
  save_models = FALSE
)
}
\arguments{
\item{data}{Data set.}

\item{vars}{String indicating the variables included in the model.}

\item{dayvar}{String indicating the day.}

\item{beepvar}{String indicating the beep.}

\item{idvar}{String indicating the subject id variable name.}

\item{estimator}{The estimator to be used. "ML" for maximum likelihood estimation
or "FIML", for full-information maximum likelihood estimation. By default set to ML}

\item{network_type}{Type of network to estimate. Currently implemented are "saturated" or "pruned".
By default set to "saturated". When "pruned" is specified, non significant edges are pruned from the model with an alpha rate of 0.05.}

\item{homogeneity_test}{Type of homogeneity test. "homogeneity_overall", refers to
test both contemporaneous and temporal network structures for homogeneity,
"homogeneity_contemporaneous" to test contemporaneous network structures for homogeneity,
"homogeneity_temporal" to test temporal network structures for homogeneity.
By default set to "homogeneity_overall".}

\item{save_models}{If TRUE, all models will be saved. By default set to FALSE to save memory.}
}
\description{
Function to test for homogeneity between idiographic network models
}
