% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INIT.R
\name{INIT}
\alias{INIT}
\title{Function to test homogeneity in idiograhic network models}
\usage{
INIT(
  data,
  vars,
  beepvar = NULL,
  dayvar = NULL,
  idvar,
  estimator = "ML",
  test = "homogeneity",
  networkType = "saturated",
  saveModels = FALSE,
  progressbar = TRUE
)
}
\arguments{
\item{data}{dataset}

\item{vars}{string indicating the variables included in the model}

\item{beepvar}{string indicating the beep}

\item{dayvar}{string indicating the day}

\item{idvar}{string indicating the subject id variable name}

\item{estimator}{type of estimator, i.e., "ML" or "FIML". By default set to ML}

\item{test}{type of homogeneity test, i.e., "homogeneity", "homogeneity_contemp", "homogeneity_temp". By default set to homogeneity}

\item{networkType}{type of network, i.e., saturated or pruned network model. By default set to saturated}

\item{saveModels}{if TRUE, all models will be saved. By default set to FALSE to save memory}

\item{progressBar}{if TRUE progress bar is included}
}
\description{
Function to test homogeneity in idiograhic network models
}
