% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INIT.R
\name{INIT}
\alias{INIT}
\title{Function to test for homogeneity between idiographic network models}
\usage{
INIT(
  data,
  vars,
  dayvar,
  beepvar,
  idvar,
  estimator = "ML",
  network_type = "saturated",
  alpha = 0.05,
  homogeneity_test = "overall",
  save_models = FALSE,
  ...
)
}
\arguments{
\item{data}{Data set.}

\item{vars}{String indicating the variables included in the model.}

\item{dayvar}{String indicating the day.}

\item{beepvar}{String indicating the beep.}

\item{idvar}{String indicating the subject id variable name.}

\item{estimator}{Estimator to be used defaults to "ML". "FIML" for full-information maximum likelihood estimation can be used, as well as any of the estimators implemented in psychonetrics.}

\item{network_type}{Type of network to estimate. Currently implemented are "saturated" or "pruned".
By default set to "saturated". When "pruned" is specified, non significant edges are pruned from the model.}

\item{alpha}{Alpha level under which significant edges are pruned from the network model. Default is set to 0.05.}

\item{homogeneity_test}{Type of homogeneity test. "overall", refers to
test both contemporaneous and temporal network structures for homogeneity,
"contemporaneous" to test contemporaneous network structures for homogeneity,
"temporal" to test temporal network structures for homogeneity.
By default set to "overall".}

\item{save_models}{If TRUE, model output for contemporaneous and temporal network models will be saved and returned as output.
By default set to FALSE to save memory.}
}
\description{
Function to test for homogeneity between idiographic network models
}
